# swagger.yml
openapi: 3.0.3
info:
  title: TP Nest Web Socket API
  version: 1.0.0
  description: Simple Swagger spec for currently implemented routes.

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: App
  - name: Auth
  - name: Profile

paths:
  /:
    get:
      tags: [App]
      summary: Health/check route
      responses:
        '200':
          description: Simple welcome message
          content:
            text/plain:
              schema:
                type: string
                example: Hello World!

  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation or duplicate data error

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: Auth success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials

  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current token
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Logout success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Missing or invalid token

  /profile:
    get:
      tags: [Profile]
      summary: Get current authenticated profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile from JWT payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          example: louis
        email:
          type: string
          format: email
          example: louis@example.com
        password:
          type: string
          example: myStrongPassword123

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: louis@example.com
        password:
          type: string
          example: myStrongPassword123

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Logged out successfully

    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: louis
        email:
          type: string
          format: email
          example: louis@example.com
        password:
          type: string
          example: $2b$10$...

    ProfileResponse:
      type: object
      properties:
        userId:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: louis@example.com